tensorflow/contrib/tensorbox/requirements.txt
runcython>=0.2.5
opencv-python>=3.2
scipy>=0.15.1
